\name{crmp-package}
\alias{crmp-package}
\alias{crmp}
\docType{package}
\title{crmp-package
}
\description{
  This package contains a variety of code which is used to interface with and process data from the Climate Related Monitoring Program (CRMP). It includes code for reading and parsing data from data files received from BC ministries and agencies, code for calculating station attributes and descriptive quantities, and code for interacting with PCIC's database containing the Provincial Climate Data Set (PCDS).
}
\details{
\tabular{ll}{
Package: \tab crmp\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-05-15\cr
License: \tab GPL\cr
Depends: \tab methods\cr
}

}
\author{
James Hiebert <hiebert@uvic.ca>
}

\keyword{ package }
