\name{BCH.xform}
\alias{BCH.xform}
\alias{EC_Buoy.xform}
\alias{FRBC.xform}
\alias{FRBCClimate.xform}
\alias{MoAg.xform}
\alias{MoE.xform}
\alias{MoE_AIRPROD1.xform}
\alias{MoE_AIRPROD2.xform}
\alias{MoFR.xform}
\alias{MoFR_research.xform}
\alias{MoFR_research_II.xform}
\alias{MoTI.xform}
\alias{MoTIe.xform}
\alias{MoTIm.xform}
\alias{RTA.xform}

\docType{data}

\title{Variable definition tables for PCDS data sets}
\description{
  Each of the networks in the Provincial Climate Data Set record and
  report meteorological quantities (i.e. 'variables') in different ways.
  All of the *.xform tables describe the network variables and how they
  should map to CF compliant quantities and units.
}
\format{CSV, sep=';'}
\keyword{datasets}
