\name{thresholds}

\docType{data}

\title{Threshold tables for QC threshold filters}
\description{
  This data table contains the minimum and maximum values for which a
  given variable is valid.  Rows are labeled with the CF standard name
  and columns 'min.valid' and 'max.valid' are available.
}
\format{Rdata}
\keyword{datasets}
